requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.2
pycodestyle==2.6.0
pymongo==3.11.3
PyYAML==5.3.1
requests==2.25.1
selenium==3.141.0
soupsieve==2.1
toml==0.10.2
urllib3==1.26.2
